requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==1.10.0
